Applied maven standard directory layout
