+++ src/main/java/nayuki/arithcode/AdaptiveArithmeticCompress.java
