+++ src/main/java/nayuki/arithcode/ArithmeticCoderBase.java
