+++ src/main/java/nayuki/arithcode/ArithmeticCompress.java
