+++ src/main/java/nayuki/arithcode/ArithmeticDecoder.java
