+++ src/main/java/nayuki/arithcode/ArithmeticEncoder.java
