+++ src/main/java/nayuki/arithcode/BitInputStream.java
