+++ src/main/java/nayuki/arithcode/BitOutputStream.java
