+++ src/main/java/nayuki/arithcode/CheckedFrequencyTable.java
