+++ src/main/java/nayuki/arithcode/FlatFrequencyTable.java
