+++ src/main/java/nayuki/arithcode/FrequencyTable.java
