+++ src/main/java/nayuki/arithcode/SimpleFrequencyTable.java
