+++ src/test/java/nayuki/arithcode/AdaptiveArithmeticCompressTest.java
