+++ src/test/java/nayuki/arithcode/ArithmeticCompressTest.java
